doc: add conditional example for setBreakpoint()
The `node-inspect` debugging client supports passing an optional third
parameter as a string to be evaluated when the breakpoint is hit. If the
condition evaluates to `true` in the current context, the breakpoint
pauses execution; otherwise the execution continues.

This was raised as an issue in
https://github.com/nodejs/node-inspect/issues/68, but the client already
supports that functionality, so I thought it'd be helpful to add it to
the node documentation.

PR-URL: https://github.com/nodejs/node/pull/35823
Reviewed-By: Rich Trott <rtrott@gmail.com>
Reviewed-By: Jan Krems <jan.krems@gmail.com>