doc: add RaisinTen as a triager
I would like to apply for the role of a triager in this project.
My motivation to become a triager is to help with the issues in
this repo and the help repo, as well as learn deeper internals
of Node.js, and to eventually become a collaborator!
I hearby declare that I have read and understood the Code of
Conduct of the project and will adhere to that.

PR-URL: https://github.com/nodejs/node/pull/36404
Reviewed-By: Gireesh Punathil <gpunathi@in.ibm.com>
Reviewed-By: Rich Trott <rtrott@gmail.com>