tools: use bundled npm in update scripts
The scripts `./tools/update-babel-eslint.sh` and
`./tools/update-eslint.sh` are relying on the version of `npm` found in
the local-defined `$PATH` env.

This changeset proposes to modify these scripts to run the version of
npm bundled in the current branch (found at `./deps/npm`) - in order to:

a) Standardize the version of npm that should be use to install these
deps, avoids the pitfall of having an inadverted user run these
scripts with an unsupported/incompatible npm version.
b) Given that npm7 has a different install algorithm than npm6 that
takes into account and install peer dependencies, it might be a safer
choice to ensure what version of npm should be use during this
transitional period in which users might still have npm6 by default in
their local system.
c) Avoids the possible extra churn of having different resulting files
being shuffled around between installs due to usage of a disparate
version of the npm cli.

PR-URL: https://github.com/nodejs/node/pull/37613
Reviewed-By: Rich Trott <rtrott@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Richard Lau <rlau@redhat.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Benjamin Gruenbaum <benjamingr@gmail.com>
Reviewed-By: Antoine du Hamel <duhamelantoine1995@gmail.com>