doc: add marsonya as a triager
I would like to apply for the role of a triager
in this project. My motivation to become a triager
is to help manage issues, as well as learn deeper
concepts of Node.js, and eventually become a
collaborator! I hereby declare that I have read and
understood the Code of Conduct and will adhere to that.

PR-URL: https://github.com/nodejs/node/pull/37667
Reviewed-By: Rich Trott <rtrott@gmail.com>
Reviewed-By: Gireesh Punathil <gpunathi@in.ibm.com>
Reviewed-By: Michael Dawson <midawson@redhat.com>
Reviewed-By: Darshan Sen <raisinten@gmail.com>
Reviewed-By: Pooja D P <Pooja.D.P@ibm.com>